+++ src/api/java/blusunrize/immersiveengineering/api/fluid/IPressurizedFluidOutput.java
@@ -1,0 +1,11 @@
+package blusunrize.immersiveengineering.api.fluid;
+
+import net.minecraftforge.fluids.FluidStack;
+
+public interface IPressurizedFluidOutput
+{
+	default int getMaxAcceptedFluidAmount(FluidStack resource)
+	{
+		return IFluidPipe.AMOUNT_PRESSURIZED;
+	}
+}