+++ src/main/java/blusunrize/immersiveengineering/common/blocks/metal/FluidPipeTileEntity.java
@@ -13,6 +13,7 @@ import blusunrize.immersiveengineering.api.IEProperties.VisibilityList;
 import blusunrize.immersiveengineering.api.IETags;
 import blusunrize.immersiveengineering.api.Lib;
 import blusunrize.immersiveengineering.api.fluid.IFluidPipe;
+import blusunrize.immersiveengineering.api.fluid.IPressurizedFluidOutput;
 import blusunrize.immersiveengineering.api.utils.CapabilityReference;
 import blusunrize.immersiveengineering.api.utils.SafeChunkUtils;
 import blusunrize.immersiveengineering.api.utils.shapes.CachedVoxelShapes;
@@ -475,6 +476,10 @@ public class FluidPipeTileEntity extends IEBaseTileEntity implements IFluidPipe,
 
 		private int getTransferableAmount(FluidStack resource, BlockEntity target)
 		{
+			// certain targets may override the transfer limits. This only works for direct pipe connections.
+			if(target instanceof IPressurizedFluidOutput pressurizedOutput)
+				return pressurizedOutput.getMaxAcceptedFluidAmount(resource);
+
 			return IFluidPipe.getTransferableAmount(
 					(resource.hasTag()&&resource.getOrCreateTag().contains(IFluidPipe.NBT_PRESSURIZED))
 							||pipe.canOutputPressurized(target, false)