+++ changelog.md
@@ -1,4 +1,5 @@
 ##### Version 1.16.5-5.0.8-pre
+- Add a new API interface that allows pressurized fluid in- and output above normal levels (BluSunrize)
 - Fix flipped bounding boxes for the Lightning Rod (BluSunrize)
 - Fix texture glitches on Refinery model (BluSunrize)
 - Fix pipes and pumps transferring fluids much slower than planned, because pressurization didn't work (BluSunrize)